% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{create_memory}
\alias{create_memory}
\title{Create Memory}
\usage{
create_memory(num_words = .words, num_bits = .bits)
}
\arguments{
\item{num_words}{number of words}

\item{num_bits}{number of bits in a word - must be a multiple of 8}
}
\value{
a list of binaryLogic::binary elements, initialized to 0
}
\description{
create a list of binaryLogic::binary elements
}

