% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{decode_fraction}
\alias{decode_fraction}
\title{Decode Fraction}
\usage{
decode_fraction(value)
}
\arguments{
\item{value}{a numeric representing a scaled fraction}
}
\value{
the R numeric equivalent assuming the word is a signed fraction
}
\description{
Convert an ILLIAC fraction to an R numeric
}

