% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{decode_integer}
\alias{decode_integer}
\title{Decode Integer}
\usage{
decode_integer(word)
}
\arguments{
\item{word}{a double representing an ILLIAC word as an integer}
}
\value{
the input word
}
\description{
ILLIAC words are stored in the emulator as integers in a
double, since R integers are on 32 bits. As long as nothing creates a non-
integer we can just treat these as words.
}

