% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{emulate}
\alias{emulate}
\title{Emulate}
\usage{
emulate(a, q, program_counter, memory)
}
\arguments{
\item{a}{the accumulator register}

\item{q}{the quotient register}

\item{program_counter}{2 * the actual program counter + 0 for left or 1 for right}

\item{memory}{the memory}
}
\value{
a list containing
\itemize{
\item display a data frame from "display_state" with the terminal state
\item a the accumulation
\item q the quotient
\item program_counter the coded program counter
\item memory the memory}
}
\description{
The emulator itself. The user calls it with an initial state.
It proceeds order by order until a stopping condition happens. At termination
the emulator returns the reason for stopping and the state at termination.
}

