% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{encode_fraction}
\alias{encode_fraction}
\title{Encode Fraction}
\usage{
encode_fraction(value)
}
\arguments{
\item{value}{numeric the number to encode}
}
\value{
the binaryLogic::binary equivalent. Note that the bounds are silenty enforced
}
\description{
Convert an R numeric to an ILLIAC signed fraction
}

