% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{encode_integer}
\alias{encode_integer}
\title{Encode Integer}
\usage{
encode_integer(value)
}
\arguments{
\item{value}{numeric the number to encode}
}
\value{
the binaryLogic::binary equivalent. Note that any fractional part is
discarded and the bounds are silenty enforced
}
\description{
Convert an R numeric as an ILLIAC signed integer
}

