% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{hex_digits}
\alias{hex_digits}
\title{Hexadecimal Digits}
\usage{
hex_digits()
}
\value{
a character vector containing the ILLIAC hexadecimal digits
}
\description{
Most ILLIAC docs use the term "sexadecimal" for base 16.
Moreover, the digits greater than 9 aren't the ones we normally use, but
"K", "S", "N", "J", "F", and "L". which had the right bit patterns on the
five-hole paper tape ILLIAC used.
}
\examples{
digits <- hex_digits()
}

