% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{illiac_add}
\alias{illiac_add}
\title{ILLIAC Add}
\usage{
illiac_add(augend, addend)
}
\arguments{
\item{augend}{the number to be augmented}

\item{addend}{the number to be added to the augend}
}
\value{
the sum of the augend and the addend
}
\description{
ILLIAC addition is a straightforward binary addition of
two's-complement numbers.
}

