% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{illiac_aq}
\alias{illiac_aq}
\title{ILLIAC AQ Doubleword}
\usage{
illiac_aq()
}
\value{
an 80-element vector with named elements
}
\description{
The ILLIAC arithmetic unit had two main 40-bit registers, A (for
accumulator) and Q (for quotient). So we create an 80-element vector.
}
\examples{
aq <- illiac_aq()
}

