% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{illiac_aq_lshift}
\alias{illiac_aq_lshift}
\title{ILLIAC AQ Left Shift one place}
\usage{
illiac_aq_lshift(illiac_aq)
}
\arguments{
\item{illiac_aq}{the number (sign and 78 fraction bits) to be shifted}
}
\value{
the one-place-left-shift of illiac_aq
}
\description{
An ILLIAC left shift
}
\examples{
aq <- illiac_aq()
shift <- illiac_aq_lshift(aq)
}

