% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{illiac_aq_rshift}
\alias{illiac_aq_rshift}
\title{ILLIAC AQ Right Shift one place}
\usage{
illiac_aq_rshift(illiac_aq)
}
\arguments{
\item{illiac_aq}{the number (sign and 78 fraction bits) to be shifted}
}
\value{
the one-place-right-shift of illiac_aq
}
\description{
An ILLIAC right shift
}
\examples{
aq <- illiac_aq()
shift <- illiac_aq_rshift(aq)
}

