% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{illiac_divide}
\alias{illiac_divide}
\title{ILLIAC divide}
\usage{
illiac_divide(a, q, r3)
}
\arguments{
\item{a}{binaryLogic::binary the accumulator register (upper dividend)}

\item{q}{binaryLogic::binary the quotient register (lower dividend)}

\item{binaryLogic::binary}{r3 the r3 register (divisor)}
}
\value{
a list with a and q = residue and quotient
}
\description{
Emulate an ILLIAC divide
}
\examples{
\dontrun{
r3 <- as.binary(encode_fraction(1/2), size = 5, signed = TRUE)
print(r3)
q <- r3
print(q)
a <- as.binary(encode_fraction(0), size = 5, signed = TRUE)
print(a)
product <- illiac_multiply(a, q, r3)
print(product$a)
print(product$q)
division <- illiac_divide(product$a, product$q, r3)
print(division$a)
print(division$q)}
}

