% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{illiac_hex_encode}
\alias{illiac_hex_encode}
\title{ILLIAC hexadecimal encode}
\usage{
illiac_hex_encode(digit)
}
\arguments{
\item{digit}{an integer from 0 to 15}
}
\value{
the ILLIAC sexadecimal code for that digit
}
\description{
Most ILLIAC docs use the term "sexadecimal" for base 16.
Moreover, the digits greater than 9 aren't the ones we normally use, but
"K", "S", "N", "J", "F", and "L". which had the right bit patterns on the
five-hole paper tape ILLIAC used.
}

