% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{illiac_multiply}
\alias{illiac_multiply}
\title{ILLIAC multiply}
\usage{
illiac_multiply(a, q, r3)
}
\arguments{
\item{a}{binaryLogic::binary the accumulator register}

\item{q}{binaryLogic::binary the quotient register (multiplier)}

\item{binaryLogic::binary}{r3 the r3 register (multiplicand)}
}
\value{
a list with a and q = 78-bit product
}
\description{
Emulate an ILLIAC multiply
}
\examples{
\dontrun{
r3 <- as.binary(encode_fraction(1/2), size = 5, signed = TRUE)
print(r3)
q <- r3
print(q)
a <- as.binary(encode_fraction(0), size = 5, signed = TRUE)
print(a)
product <- illiac_multiply(a, q, r3)
print(product$a)
print(product$q)}
}

