% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{illiac_word}
\alias{illiac_word}
\title{ILLIAC word}
\usage{
illiac_word()
}
\value{
a 40-element vector with named elements
}
\description{
ILLIAC used a 40-bit word. There are numerous ways we could deal
with this, but since ILLIAC's "native" arithmetic is based on two's-complement
fractions, there's nothing native to R that deals with it easily.

So we might as well be pedagocial and just use a 40-element vector of
individual bits. And as a bonus we can name them and use matrix and vector
operations on them.
}
\examples{
accumulator <- illiac_word()
quotient <- illiac_word()
}

