% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{illiac_word_length}
\alias{illiac_word_length}
\title{ILLIAC Word Length}
\usage{
illiac_word_length()
}
\value{
the length of an ILLIAC word (40 bits)
}
\description{
Number of bits in an ILLIAC word - 40
}
\examples{
N <- illiac_word_length()
}

