% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{interpret_as_illiac}
\alias{interpret_as_illiac}
\title{Interpret as ILLIAC}
\usage{
interpret_as_illiac(illiac_words)
}
\arguments{
\item{illiac_words}{a vector of ILLIAC words, stored as integers in doubles}
}
\value{
a list of
\itemize{
\item fraction double - the input vector scaled to print as fractions
\item left_order integer the left order
\item left_address integer the left 12-bit address
\item right_order integer the right order
\item right_address integer the left 12-bit address}
}
\description{
Interpret a vector of ILLIAC words in doubles as fractions and order pairs
}

