% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{order_code_table}
\alias{order_code_table}
\title{Order Code Table}
\usage{
order_code_table()
}
\value{
an order code table - a matrix with the type and variant digits as
row and column names
}
\description{
create and return empty order code table
}
\examples{
octable <- order_code_table()
}

