% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{sadoi}
\alias{sadoi}
\title{SADOI}
\usage{
sadoi(source_file)
}
\arguments{
\item{source_file}{path to a file with directives and orders in SADOI format}
}
\value{
a list of
\itemize{
\item memory a "memory" object containing the assembled orders
\item start the program counter for the last start directive encountered}
}
\description{
A re-implementation of the ILLIAC Symbolic Address Decimal
Order Input (SADOI) assembler.
}

