% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{create_memory}
\alias{create_memory}
\title{Create Memory}
\usage{
create_memory(num_words = .words)
}
\arguments{
\item{num_words}{number of words}
}
\value{
a numeric vector initialized to 0
}
\description{
create a vector of integers
}

