% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{create_memory}
\alias{create_memory}
\title{Create Memory}
\usage{
create_memory(num_words, num_bits)
}
\arguments{
\item{num_words}{number of rows (words, bytes, etc.)}

\item{num_bits}{number of bits in a word - must be a multiple of 8}
}
\value{
a num_words by num_bits array filled with FALSE
}
\description{
create a boolean matrix representing a computer memory. The rows
are words (addressable entities) and the columns are bits.
}

