% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{decode_fraction}
\alias{decode_fraction}
\title{Decode Fraction}
\usage{
decode_fraction(binary)
}
\arguments{
\item{binary}{a binaryLogic::binary ILLIAC word}
}
\value{
the R numeric equivalent assuming the word is a signed fraction
}
\description{
Convert an ILLIAC fraction to an R numeric
}

