% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{decode_integer}
\alias{decode_integer}
\title{Decode Integer}
\usage{
decode_integer(binary)
}
\arguments{
\item{binary}{a binaryLogic::binary ILLIAC word}
}
\value{
the R numeric equivalent assuming the word is a signed integer
}
\description{
Convert an ILLIAC signed integer to an R numeric
}

