% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emulator.R
\name{encode_integer}
\alias{encode_integer}
\title{Encode Integer}
\usage{
encode_integer(value)
}
\arguments{
\item{value}{numeric the number to encode}
}
\value{
a double truncated to ILLIAC signed integer range
}
\description{
We store ILLIAC words as integers in a double. So all we need to
do to encode an integer is enforce bound.
}

